requirements.txt
malmoenv>=0.0.8
gymnasium>=1.0.0
numpy>=1.20.0
pyyaml>=5.4.0
pillow>=8.0.0
